backend/requirements.txt
fastapi
uvicorn
pytest
httpx
psycopg[binary]
python-dotenv
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
python-multipart
pydantic[email]
google-genai